+++ NEWS.md
@@ -1,3 +1,6 @@
+# TableTraits.jl v0.1.0 Release Notes
+* Add get_columns_copy interface
+
 # TableTraits.jl v0.0.3 Release Notes
 * Use proper precompile declaration